fix: Address PR review issues
- Fix 2 failing tests in useHtmlSlots hook test suite
- Add debouncing to MutationObserver callbacks for better performance (300ms delay)
- Split large SlotManager component into smaller, focused components:
  - HtmlSlotCard: Individual HTML slot management component
  - HtmlSlotDetectionPanel: Slot detection status and controls
  - WidgetCard: Widget display and action components
- Reduce SlotManager from 1135 lines to ~400 lines (64% reduction)
- Improve code maintainability and modularity
- All 18 tests now passing
